chore(CreateTearsheet): extract more logic into custom hooks and utils
