+++ packages/cloud-cognitive/src/components/CreateFullPage/CreateFullPage.js
@@ -28,29 +28,11 @@ import { CreateInfluencer } from '../CreateInfluencer';
 import { ActionSet } from '../ActionSet';
 import { usePreviousValue } from '../../global/js/use/usePreviousValue';
 import { useValidCreateStepCount } from '../../global/js/use/useValidCreateStepCount';
-import { useCreateComponentStepChange } from '../../global/js/use/useCreateComponentStepChange';
+import { useCreateComponentFocus } from '../../global/js/use/useCreateComponentFocus';
+import { isValidChildren } from '../CreateTearsheet/utils/isValidChildren';
 
 const blockClass = `${pkg.prefix}--create-full-page`;
 const componentName = 'CreateFullPage';
-
-// Custom PropType validator which checks and ensures that the children property has no more than 4 nodes
-const isValidChildren =
-  () =>
-  ({ children }) => {
-    children.length > 1 &&
-      children.map((child) => {
-        if (
-          child &&
-          child.props &&
-          child.props.type !== CREATE_FULL_PAGE_STEP
-        ) {
-          throw new Error(
-            'Each child of Create Full Page is required to be a `CreateFullPageStep`. Please remove the HTML element, or wrap it around the `CreateFullPageStep` component.'
-          );
-        }
-        return;
-      });
-  };
 
 export let CreateFullPage = React.forwardRef(
   (
@@ -100,7 +82,7 @@ export let CreateFullPage = React.forwardRef(
       return steps;
     }, [children]);
 
-    useCreateComponentStepChange(
+    useCreateComponentFocus(
       previousState,
       currentStep,
       getFullPageSteps,
@@ -432,44 +414,6 @@ export let CreateFullPage = React.forwardRef(
       );
     };
 
-    // track scrolling/intersection of create sections so that we know
-    // which section is active (updates the SideNavItems `isActive` prop)
-    useEffect(() => {
-      if (shouldViewAll) {
-        const fullPageMainContent = document.querySelector(
-          `.${blockClass}__content`
-        );
-        let options = {
-          root: fullPageMainContent,
-          rootMargin: '0px',
-          threshold: 0,
-        };
-        // Convert NodeList to array so we can find the index
-        // of the section that should be marked as `active`.
-        const viewAllSections = Array.from(
-          document.querySelectorAll(
-            `.${blockClass}__section.${blockClass}__step--visible-section`
-          )
-        );
-        const observer = new IntersectionObserver((entries) => {
-          // isIntersecting is true when element and viewport/options.root are overlapping
-          // isIntersecting is false when element and viewport/options.root don't overlap
-          if (entries[0].isIntersecting) {
-            // DOM element that is intersecting
-            const visibleTarget = entries[0].target;
-            // Get visible element index
-            const visibleTargetIndex = viewAllSections.findIndex(
-              (item) => item.id === visibleTarget.id
-            );
-            setActiveSectionIndex(visibleTargetIndex);
-          }
-        }, options);
-        viewAllSections.forEach((section) => {
-          observer.observe(section);
-        });
-      }
-    }, [shouldViewAll]);
-
     return (
       <div {...rest} ref={ref} className={cx(blockClass, className)}>
         <div className={`${blockClass}__influencer`}>
@@ -482,6 +426,8 @@ export let CreateFullPage = React.forwardRef(
             includeViewAllToggle={includeViewAllToggle}
             handleToggleState={(toggleState) => setShouldViewAll(toggleState)}
             handleActiveSectionIndex={(index) => setActiveSectionIndex(index)}
+            open={open}
+            previousState={previousState}
             sideNavAriaLabel={sideNavAriaLabel}
             toggleState={shouldViewAll}
             viewAllToggleLabelText={viewAllToggleLabelText}
@@ -553,7 +499,10 @@ CreateFullPage.propTypes = {
   /**
    * The main content of the full page
    */
-  children: isValidChildren(),
+  children: isValidChildren({
+    componentName,
+    stepType: CREATE_FULL_PAGE_STEP,
+  }),
 
   /**
    * Provide an optional class to be applied to the containing node.
