+++ packages/cloud-cognitive/src/components/CreateTearsheet/CreateTearsheet.js
@@ -20,34 +20,18 @@ import { extractShapesArray } from '../../global/js/utils/props-helper';
 import wrapFocus from '../../global/js/utils/wrapFocus';
 import { TearsheetShell } from '../Tearsheet/TearsheetShell';
 import { CreateInfluencer } from '../CreateInfluencer';
+import { CreateTearsheetDivider } from './CreateTearsheetDivider';
 import { carbon, pkg } from '../../settings';
 import { CREATE_TEARSHEET_SECTION, CREATE_TEARSHEET_STEP } from './constants';
 import { usePreviousValue } from '../../global/js/use/usePreviousValue';
 import { useValidCreateStepCount } from '../../global/js/use/useValidCreateStepCount';
 import { useResetCreateComponent } from '../../global/js/use/useResetCreateComponent';
+import { useCreateComponentFocus } from '../../global/js/use/useCreateComponentFocus';
 import { useCreateComponentStepChange } from '../../global/js/use/useCreateComponentStepChange';
+import { isValidChildren } from './utils/isValidChildren';
 
 const componentName = 'CreateTearsheet';
 const blockClass = `${pkg.prefix}--tearsheet-create`;
-
-// Custom PropType validator which checks and ensures that the children property has no more than 4 nodes
-const isValidChildren =
-  () =>
-  ({ children }) => {
-    children.length > 1 &&
-      children.map((child) => {
-        if (
-          child &&
-          child.props &&
-          child.props.type !== CREATE_TEARSHEET_STEP
-        ) {
-          throw new Error(
-            `Each child of ${componentName} is required to be a ${CREATE_TEARSHEET_STEP}. Please remove the HTML element, or wrap it around the ${CREATE_TEARSHEET_STEP} component.`
-          );
-        }
-        return;
-      });
-  };
 
 export let CreateTearsheet = forwardRef(
   (
@@ -98,7 +82,7 @@ export let CreateTearsheet = forwardRef(
       return steps;
     }, [children]);
 
-    useCreateComponentStepChange(
+    useCreateComponentFocus(
       previousState,
       currentStep,
       getTearsheetSteps,
@@ -106,6 +90,24 @@ export let CreateTearsheet = forwardRef(
     );
     useValidCreateStepCount(getTearsheetSteps, componentName);
     useResetCreateComponent(previousState, open, setCurrentStep);
+    useCreateComponentStepChange({
+      setCurrentStep,
+      setIsSubmitting,
+      setShouldViewAll,
+      onClose,
+      onRequestSubmit,
+      componentName,
+      getComponentSteps: getTearsheetSteps,
+      currentStep,
+      shouldViewAll,
+      backButtonText,
+      cancelButtonText,
+      submitButtonText,
+      nextButtonText,
+      isSubmitting,
+      componentBlockClass: blockClass,
+      setCreateComponentActions: setCreateTearsheetActions,
+    });
 
     // Log a warning to the console in the event there are no CreateTearsheetSection components
     // inside of the CreateTearsheetSteps when the viewAll toggle is provided and turned on.
@@ -175,131 +177,6 @@ export let CreateTearsheet = forwardRef(
         });
       }
     }, [includeViewAllToggle, shouldViewAll, children]);
-
-    // useEffect to handle multi step logic
-    useEffect(() => {
-      const onUnmount = () => {
-        setCurrentStep(0);
-        setIsSubmitting(false);
-        setShouldViewAll(false);
-        onClose();
-      };
-      const handleOnRequestSubmit = async () => {
-        // check if onRequestSubmit returns a promise
-        try {
-          await onRequestSubmit();
-          onUnmount();
-        } catch (error) {
-          setIsSubmitting(false);
-          console.warn(`${componentName} submit error: ${error}`);
-        }
-      };
-      const isSubmitDisabled = () => {
-        let step = 0;
-        let submitDisabled = false;
-        let viewAllSubmitDisabled = false;
-        const tearsheetSteps = getTearsheetSteps();
-        tearsheetSteps.forEach((child) => {
-          step++;
-          if (currentStep === step) {
-            submitDisabled = child.props.disableSubmit;
-          }
-          if (shouldViewAll && child.props.disableSubmit) {
-            viewAllSubmitDisabled = true;
-          }
-        });
-        if (!shouldViewAll) {
-          return submitDisabled;
-        }
-        return viewAllSubmitDisabled;
-      };
-      const handleNext = async () => {
-        setIsSubmitting(true);
-        const createSteps = getTearsheetSteps();
-        if (createSteps[currentStep - 1].props.onNext) {
-          try {
-            await createSteps[currentStep - 1].props.onNext();
-            continueToNextStep();
-          } catch (error) {
-            setIsSubmitting(false);
-            console.warn(`${componentName} onNext error: ${error}`);
-          }
-        } else {
-          continueToNextStep();
-        }
-      };
-      const handleSubmit = async () => {
-        setIsSubmitting(true);
-        const createSteps = getTearsheetSteps();
-        // last step should have onNext as well
-        if (createSteps[currentStep - 1].props.onNext) {
-          try {
-            await createSteps[currentStep - 1].props.onNext();
-            await handleOnRequestSubmit();
-          } catch (error) {
-            setIsSubmitting(false);
-            console.warn(`${componentName} onNext error: ${error}`);
-          }
-        } else {
-          await handleOnRequestSubmit();
-        }
-      };
-      if (getTearsheetSteps()?.length) {
-        const createSteps = getTearsheetSteps();
-        const total = createSteps.length;
-        const buttons = [];
-        if (total > 1 && !shouldViewAll) {
-          buttons.push({
-            key: 'create-tearsheet-action-button-back',
-            label: backButtonText,
-            onClick: () => setCurrentStep((prev) => prev - 1),
-            kind: 'secondary',
-            disabled: currentStep === 1,
-          });
-        }
-        buttons.push({
-          key: 'create-tearsheet-action-button-cancel',
-          label: cancelButtonText,
-          onClick: onUnmount,
-          kind: 'ghost',
-        });
-        buttons.push({
-          key: 'create-tearsheet-action-button-submit',
-          label: shouldViewAll
-            ? submitButtonText
-            : currentStep < total
-            ? nextButtonText
-            : submitButtonText,
-          onClick: shouldViewAll
-            ? handleSubmit
-            : currentStep < total
-            ? handleNext
-            : handleSubmit,
-          disabled: isSubmitDisabled(),
-          kind: 'primary',
-          loading: isSubmitting,
-          className: `${blockClass}__create-button`,
-        });
-        setCreateTearsheetActions(buttons);
-      }
-    }, [
-      getTearsheetSteps,
-      children,
-      backButtonText,
-      cancelButtonText,
-      currentStep,
-      onClose,
-      nextButtonText,
-      submitButtonText,
-      onRequestSubmit,
-      isSubmitting,
-      shouldViewAll,
-    ]);
-
-    const continueToNextStep = () => {
-      setIsSubmitting(false);
-      setCurrentStep((prev) => prev + 1);
-    };
 
     // check if child is a tearsheet step component
     const isTearsheetStep = (child) => {
@@ -440,7 +317,7 @@ export let CreateTearsheet = forwardRef(
               <>
                 {child}
                 {shouldViewAll && !isLastSectionOfLastStep && (
-                  <span className={`${blockClass}__section--divider`} />
+                  <CreateTearsheetDivider />
                 )}
               </>
             );
@@ -487,45 +364,6 @@ export let CreateTearsheet = forwardRef(
       );
     };
 
-    // track scrolling/intersection of create sections so that we know
-    // which section is active (updates the SideNavItems `isActive` prop)
-    useEffect(() => {
-      if (shouldViewAll) {
-        const tearsheetMainContent = document.querySelector(
-          `.${pkg.prefix}--tearsheet__content`
-        );
-        let options = {
-          root: tearsheetMainContent,
-          rootMargin: '0px',
-          threshold: 0,
-        };
-        // Convert NodeList to array so we can find the index
-        // of the section that should be marked as `active`.
-        const viewAllSections = Array.from(
-          document.querySelectorAll(
-            `.${pkg.prefix}--tearsheet-create__section.${pkg.prefix}--tearsheet-create__step--visible-section`
-          )
-        );
-        /* istanbul ignore next */
-        const observer = new IntersectionObserver((entries) => {
-          // isIntersecting is true when element and viewport/options.root are overlapping
-          // isIntersecting is false when element and viewport/options.root don't overlap
-          if (entries[0].isIntersecting) {
-            // DOM element that is intersecting
-            const visibleTarget = entries[0].target;
-            // Get visible element index
-            const visibleTargetIndex = viewAllSections.findIndex(
-              (item) => item.id === visibleTarget.id
-            );
-            setActiveSectionIndex(visibleTargetIndex);
-          }
-        }, options);
-        viewAllSections.forEach((section) => {
-          observer.observe(section);
-        });
-      }
-    }, [shouldViewAll]);
-
     useResizeDetector({
       handleWidth: true,
       onResize: handleResize,
@@ -549,6 +387,8 @@ export let CreateTearsheet = forwardRef(
             includeViewAllToggle={includeViewAllToggle}
             handleToggleState={(toggleState) => setShouldViewAll(toggleState)}
             handleActiveSectionIndex={(index) => setActiveSectionIndex(index)}
+            open={open}
+            previousState={previousState}
             sideNavAriaLabel={sideNavAriaLabel}
             toggleState={shouldViewAll}
             viewAllToggleLabelText={viewAllToggleLabelText}
@@ -603,7 +443,10 @@ CreateTearsheet.propTypes = {
   /**
    * The main content of the tearsheet
    */
-  children: isValidChildren(),
+  children: isValidChildren({
+    componentName,
+    stepType: CREATE_TEARSHEET_STEP,
+  }),
 
   /**
    * An optional class or classes to be added to the outermost element.
