+++ packages/cloud-cognitive/src/components/CreateTearsheet/preview-components/MultiStepWithSectionsTearsheet.js
@@ -188,7 +188,7 @@ export const MultiStepWithSectionsTearsheet = ({
           <h4 className={`${componentBlockClass}__step--title`}>Meta data</h4>
           <fieldset className={`${componentBlockClass}__step--fieldset`}>
             <TextInput
-              labelText="Topic meta data"
+              labelText="Topic meta data (optional)"
               id="tearsheet-multi-step-story-text-input-multi-step-1-input-4"
               value={topicMetaData}
               placeholder="Enter topic meta data"
